services/account-created/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
google-cloud-pubsub